AI実装UI改変バグつぶし

+++ schemas.py
@@ -1,19 +1,20 @@
 from pydantic import BaseModel, ConfigDict
 from datetime import date
 from typing import Optional
+from decimal import Decimal
 
 class TentBase(BaseModel):
     name: str
     brand: Optional[str] = None
     price: Optional[int] = None
-    capacity: Optional[int] = None
-    weight_kg: Optional[float] = None
-    size_w: Optional[float] = None
-    size_d: Optional[float] = None
-    size_h: Optional[float] = None
-    pack_w: Optional[float] = None
-    pack_d: Optional[float] = None
-    pack_h: Optional[float] = None
+    capacity: Optional[Decimal] = None
+    weight_kg: Optional[Decimal] = None
+    size_w: Optional[Decimal] = None
+    size_d: Optional[Decimal] = None
+    size_h: Optional[Decimal] = None
+    pack_w: Optional[Decimal] = None
+    pack_d: Optional[Decimal] = None
+    pack_h: Optional[Decimal] = None
     material: Optional[str] = None
     purchase_date: Optional[date] = None
 
@@ -24,14 +25,14 @@ class TentUpdate(BaseModel):
     name: Optional[str] = None
     brand: Optional[str] = None
     price: Optional[int] = None
-    capacity: Optional[int] = None
-    weight_kg: Optional[float] = None
-    size_w: Optional[float] = None
-    size_d: Optional[float] = None
-    size_h: Optional[float] = None
-    pack_w: Optional[float] = None
-    pack_d: Optional[float] = None
-    pack_h: Optional[float] = None
+    capacity: Optional[Decimal] = None
+    weight_kg: Optional[Decimal] = None
+    size_w: Optional[Decimal] = None
+    size_d: Optional[Decimal] = None
+    size_h: Optional[Decimal] = None
+    pack_w: Optional[Decimal] = None
+    pack_d: Optional[Decimal] = None
+    pack_h: Optional[Decimal] = None
     material: Optional[str] = None
     purchase_date: Optional[date] = None
